ValueObservationRecorder cleanup
